% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_series}
\alias{list_series}
\title{List created series}
\usage{
list_series(theme_id = NULL, theme_title = NULL)
}
\arguments{
\item{theme_id, theme_title}{Optional parameters that can be used individually
or combined to filter the selected themes.}
}
\value{
A data.frame
}
\description{
List all series from the Atlas, regardless of whether they have data
available for download or not.
}
\examples{
list_series()
list_series(theme_id = c(42, 50))
list_series(theme_id = 35, theme_title = c("Organizações do Estado"))
}
