% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_themes}
\alias{list_themes}
\title{List available themes}
\usage{
list_themes()
}
\value{
A data.frame
}
\description{
\code{list_themes()} returns a data.frame containing all available themes and
their corresponding ids.
}
\examples{
list_themes()
}
